Manejo usuarios

+++ varwwwhtml/dcaas-app/app/Http/Controllers/UserController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use App\Enums\PermisosUsuario;
+use App\Http\Requests\BorrarUsuarioRequest;
 use App\Http\Requests\LoginUsuarioRequest;
 use App\Http\Requests\RegistrarUsuarioRequest;
 use App\Models\User;
@@ -11,6 +13,7 @@ use App\Responses\RespuestaAPI;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Log;
 use App\Http\Requests\EditarUsuarioRequest;
+use App\Facades\ManejadorPermisos;
 
 class UserController extends Controller
 {
@@ -38,18 +41,8 @@ class UserController extends Controller
     public function ver($id)
     {
         $user = User::find($id);
-        if (!$user) {
-            return RespuestaAPI::fallo(404, 'Usuario no encontrado');
-        }
-        return RespuestaAPI::exito('Usuario encontrado correctamente', $user->only([
-            'nickname',
-            'nombre',
-            'descripcion',
-            'url_foto',
-            'id',
-            'permisos',
-            'fecha_creacion'
-        ]));
+        if (!$user) return RespuestaAPI::fallo(404, 'Usuario no encontrado');
+        return RespuestaAPI::exito('Usuario encontrado correctamente', $user->only(['nickname','nombre','descripcion','url_foto','id','permisos','fecha_creacion']));
     }
 
     public function verYo(Request $request)
@@ -58,16 +51,7 @@ class UserController extends Controller
         $user = $request->user();
         //$token = $user->currentAccessToken();
         return RespuestaAPI::exito('Tus datos', [
-            'user' => $user->only([
-                'nickname',
-                'nombre',
-                'descripcion',
-                'url_foto',
-                'id',
-                'permisos',
-                'fecha_creacion',
-                'email'
-            ]),
+            'user' => $user->only(['nickname','nombre','descripcion','url_foto','id','permisos','fecha_creacion','email']),
             //'access_token' => $token,
             'token_type' => 'Bearer'
         ]);
@@ -85,19 +69,11 @@ class UserController extends Controller
         if (!$user || !Hash::check($datos['password'], $user->password)) {
             return RespuestaAPI::fallo(401, 'Credenciales incorrectas');
         }
+        if (ManejadorPermisos::todoRestringido($user)) return RespuestaAPI::fallo(401, "No tienes permisos para realizar esta acción");
         $user->tokens()->delete();
         $token = $user->createToken('login_token')->plainTextToken;
         return RespuestaAPI::exito('Login exitoso', [
-            'user' => $user->only([
-                'nickname',
-                'nombre',
-                'descripcion',
-                'url_foto',
-                'id',
-                'permisos',
-                'fecha_creacion',
-                'email'
-            ]),
+            'user' => $user->only(['nickname','nombre','descripcion','url_foto','id','permisos','fecha_creacion','email']),
             'access_token' => $token,
             'token_type' => 'Bearer'
         ]);
@@ -105,43 +81,47 @@ class UserController extends Controller
 
     public function editar(EditarUsuarioRequest $request)
     {
-
         $user = $request->user();
-
         $datos = $request->validated();
-
+        if (!ManejadorPermisos::puedeAutoeditar($user)) return RespuestaAPI::fallo(401, "No tienes permisos para realizar esta acción");
         if (!empty($datos['permisos']) || !empty($datos['publicante']) || !empty($datos['fecha_creacion']) || !empty($datos['id']) || !empty($datos['token'])) return RespuestaAPI::fallo(422, "Hay parámetros que no tienes permitido editar");
-
-        return response()->json(["user" => $user, "datos" => $datos, "request" => $request]);
-
-        $camposSensibles = ['nickname', 'email', 'password'];
-        $keysEnviadas = array_keys($datos);
-        $cambioSensible = array_intersect($keysEnviadas, $camposSensibles);
-
-        if (!empty($cambioSensible)) {
+        if (!empty($datos['nickname']) || !empty($datos['email']) || !empty($datos['password'])) {
             if (!Hash::check($request->oldPassword, $user->password)) {
                 return RespuestaAPI::fallo(401, 'La contraseña antigua es incorrecta');
             }
         }
-
         if (isset($datos['password'])) {
             $datos['password'] = Hash::make($datos['password']);
         }
-
         $user->fill($datos);
         $user->save();
-
         return RespuestaAPI::exito('Usuario actualizado correctamente', [
-            'user' => $user->only([
-                'id',
-                'nickname',
-                'nombre',
-                'email',
-                'descripcion',
-                'url_foto',
-                'permisos',
-                'fecha_creacion'
-            ])
+            'user' => $user->only(['id','nickname','nombre','email','descripcion','url_foto','permisos','fecha_creacion'])
+        ]);
+    }
+
+    public function borrar(BorrarUsuarioRequest $request) {
+        $user = $request->user();
+        $datos = $request->validated();
+        if (!$user || !Hash::check($datos['password'], $user->password)) {
+            return RespuestaAPI::fallo(401, 'Credenciales incorrectas');
+        }
+        $user->tokens()->delete();
+        $user->delete();
+
+        //TODO: borrado en cascada
+
+        return RespuestaAPI::exito("Usuario borrado correctamente", [
+            'user' => $user->only(['id','nickname','nombre','email','descripcion','url_foto','permisos','fecha_creacion'])
+        ]);
+    }
+
+    public function cerrarSesion(Request $request) {
+        $user = $request->user();
+        if (!$user) return RespuestaAPI::fallo(401, 'Credenciales incorrectas');
+        $user->tokens()->delete();
+        return RespuestaAPI::exito("Sesión cerrada correctamente", [
+            'user' => $user->only(['id','nickname','nombre','email','descripcion','url_foto','permisos','fecha_creacion'])
         ]);
     }
 }
